DOMA-1138 move UnitInfo component to common domain
